Refactor user and resume routes into separate files; update main app to include new routes
