+++ Backend/app/main.py
@@ -6,7 +6,8 @@ load_dotenv(find_dotenv(".env.local"), override=True)
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from .user_route import router as user_router
+from .routes.user_route import router as user_router
+from .routes.resume_route import router as resume_router
 
 app = FastAPI(title="JobAssistant API")
 
@@ -18,6 +19,7 @@ app.add_middleware(
 )
 
 app.include_router(user_router, prefix="/users", tags=["users"])
+app.include_router(resume_router, prefix="/users")
 
 
 @app.get("/health")
